Restructured connector module.
