+++ example_apps/kafka_connector/src/main.rs
@@ -33,7 +33,7 @@ use swimos::{
     server::{Server, ServerBuilder},
 };
 use swimos_connector::ConnectorModel;
-use swimos_connector_kafka::{KafkaConnector, KafkaIngressConfiguration};
+use swimos_connector_kafka::{KafkaIngressConfiguration, KafkaIngressConnector};
 use swimos_recon::parser::parse_recognize;
 
 mod params;
@@ -57,7 +57,7 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync>> {
     let route = RoutePattern::parse_str("/kafka")?;
 
     let connector_agent =
-        ConnectorModel::for_fn(move || KafkaConnector::for_config(connector_config.clone()));
+        ConnectorModel::for_fn(move || KafkaIngressConnector::for_config(connector_config.clone()));
 
     let server = ServerBuilder::with_plane_name("Example Plane")
         .set_bind_addr("127.0.0.1:8080".parse()?)
